+++ WebServer/WebServer.cpp
@@ -24,7 +24,7 @@ namespace Plugin {
         // Setup skip URL for right offset.
         _pid = 0;
         _service = service;
-        _skipURL = _service->WebPrefix().length();
+        _skipURL = static_cast<uint32_t>(_service->WebPrefix().length());
 
         config.FromString(_service->ConfigLine());
